Suppress deprecation notice for Symfony TableHelper.
We'll deal with it for reals when Console component hits 3.0.

Since everyone does (or should) use `E_ALL` in development, `E_USER_DEPRECATED` is disruptive enough that it's essentially a backwards compatibility break. This is not cool. My users should not have to choose between requiring an upgrade on a point release, and not getting any more bug fixes or new features from PsySH.

Fixes #132