Add Doc support for constants. Closes #384
This follows similar conventions to the existing `doc` process:

* To allow for multi-language support the constants are maintained in it's own SQLite database named `php_constants.sqlite`. Since sconstants are not _atomically_ generated as part of PHP's manual builder process these have been added by hand to the database (English only right now).

* The 'doc keyword' command will now check if the keyword is a system constant (e.g. PHP_VERSION)
or a "magic" constant (e.g. __DIR__) if either of these is true then the documentation will be
read from the `php_constants` database and presented as output similar to how language constructs or innate functions are output.

Note: Because of the lack of reflection for these constants the output makes limited use of the templating engine. I tried to "hack" the reflection helpers (like what is being done for language constructs such as `echo` but this only proved an exercise in frustration.