Remove length limit on transient readline's fgets.
This is unnecessary in PHP >= 5, and results in truncated user input for really long lines. See excellent discussion in #496 for more context.

Note that fixing this doesn't completely remove line length limits, but raises them to the maximum supported by TTY on the OS, which is probably 4096 on Linux, for example.