Move PHPUnit to a composer bin dependency
Add a `make test` target and update CONTRIBUTING instructions. Update Travis config to use `make test` on CI.

We aren't super picky about which version of PHPUnit we use, and really there is no reason to complicate our runtime dependencies by taking on all of PHPUnit's dependencies as well. So now we can use globally-installed PHPUnit, or a local phar, or a composer-bin-installed version. Basically anything but a dev dependency :)