If not completing a valid prefix, consider it an empty string
This ensures that a $token which is valid neither as an identifier nor
as a variable will only appear as a 'previous' token, and never as the
token being actively completed.  This reduces the variety of cases
which the matchers need to care about.

This also ensures that completing against whitespace does not attempt
to complete the preceding token (which was already complete, and not
what the user asked for).