Psy\debug: Option to suppress printing of context
Option for `Psy\debug` to suppress printing of context (via `whereami` command):

Background: I implemented a simple shell script to boot PsySH with our own variables and found it quite irritating - especially for other developers using that script - that parts of it were printed before getting a prompt. Would be happy if you could merge this!