Stop testing HHVM.
It is possible to test HHVM via GitHub actions, and we can look into that if anyone is really excited about it. But given that the last LTS HHVM version with PHP compatibility has now been unsupported for over a year, I'm inclined to just ... stop. We're not intentionally breaking PsySH on HHVM, nor are we removing compatibility code. But we will no longer be testing on CI.

This also ends our dependency on Travis CI. It was great while it lasted, but they've effectively stopped supporting open source projects, so the writing has been on the wall for a while. As of the time of this commit, I've spent twenty-eight minutes waiting for the final HHVM test job to queue. All other tests, for every other supported PHP version, finished at least twenty-seven minutes ago.

I'm just going to put it out of its misery.

:rip: Thanks for all the builds, Travis.