Don't use broken BusyBox less as default pager.
Fixes #778

This is the flavor of less that ships by default in Alpine Linux. We're detecting it by checking whether `less` is a symlink that includes `busybox` in the path. There may be false negatives, but shouldn't be (m)any false positives?

This default can be overridden by explicitly setting less as the pager via PsySH config, if you really want to, but you probably don't. Install proper `less` instead :)