Print shorter return values for actions than inspection.
See #512

We can adjust what exactly "shorter" looks like; in this case it's
completely collapsed—so an array would be `[…5]` or an object would just
be the class and a number—but showing one level deep might make sense.

We also can adjust the heuristics used; as a first approximation,
anything that is doing assignment or calling a method that looks like an
action gets the abbreviated return value, and anything that looks like
inspection—for example bare variable names, property access, calling
method prefixes like `find`, `is`, `get`, etc—will get a slightly _more_
verbose dump than before.